backend/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
